server/python/pyproject.toml
[tool.poetry]
name = "example_server"
version = "0.1.0"
description = ""
authors = ["Markus Ecker <markus.ecker@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ag-ui-protocol = "^0.1.5"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
fastmcp = "^2.10.6"
flask = "^3.1.1"
langchain-mcp = "^0.2.1"
langchain-openai = "^0.3.28"
langgraph = "^0.5.3"
langchain-mcp-adapters = "^0.1.9"
azure-identity = "^1.23.1"
openai = "^1.97.0"
pandas = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "example_server:main"